feat(single-import sdk): move `aruba` to `pkg` (issue #9)
