+++ cmd/example/delete.go
@@ -8,8 +8,8 @@ import (
 	"os"
 	"time"
 
-	aruba "github.com/Arubacloud/sdk-go"
 	"github.com/Arubacloud/sdk-go/internal/restclient"
+	"github.com/Arubacloud/sdk-go/pkg/aruba"
 )
 
 // runDeleteExample demonstrates how to delete all resources
