+++ cmd/example/main.go
@@ -9,8 +9,8 @@ import (
 	"os"
 	"time"
 
-	aruba "github.com/Arubacloud/sdk-go"
 	"github.com/Arubacloud/sdk-go/internal/restclient"
+	"github.com/Arubacloud/sdk-go/pkg/aruba"
 	"github.com/Arubacloud/sdk-go/pkg/types"
 )
