+++ pkg/aruba/client.go
