+++ pkg/aruba/compute.go
