+++ pkg/aruba/metric.go
