+++ pkg/aruba/storage.go
